+++ EFM8/Robot/Source_Code/inductance.h
@@ -2,7 +2,7 @@
 
 #define PERIOD_PIN P1_0 // Signal to Measure
 
-unsigned long GetPeriod (int n);
-float GetFrequency_Hz(void);
+unsigned long GetPeriod(int n);
+unsigned long GetFrequency_Hz(void);
 float GetCapacitance_mF(void);
 int GetInductance_microH(void);
